Cast int values for consistency in json tests

+++ plugins/Diagnostics/Commands/ArchivingInstanceStatistics.php
@@ -51,17 +51,17 @@ class ArchivingInstanceStatistics extends ConsoleCommand
     public function getArchivingInstanceStatistics(): array
     {
         $stats = [];
-        $stats[] = ['Site Count', Db::fetchOne("SELECT COUNT(*) FROM " . Common::prefixTable("site"))];
-        $stats[] = ['Segment Count', Db::fetchOne("SELECT COUNT(*) FROM " . Common::prefixTable("segment"))];
+        $stats[] = ['Site Count', (int) Db::fetchOne("SELECT COUNT(*) FROM " . Common::prefixTable("site"))];
+        $stats[] = ['Segment Count', (int) Db::fetchOne("SELECT COUNT(*) FROM " . Common::prefixTable("segment"))];
         $stats[] = ['Database Version', defined('PIWIK_TEST_MODE') ? 'mysql-version-redacted' : Db::get()->getServerVersion()];
-        $stats[] = ['Last full Month Hits', Db::fetchOne(
+        $stats[] = ['Last full Month Hits', (int) Db::fetchOne(
             "SELECT COUNT(*) FROM " . Common::prefixTable("log_link_visit_action") . " WHERE server_time >= ? AND server_time <= ?",
             [
                 Date::now()->setDay(1)->subMonth(1)->setTime('00:00:00')->toString('Y-m-d H:i:s'),
                 Date::now()->setDay(1)->subDay(1)->setTime('23:59:59')->toString('Y-m-d H:i:s')
             ])
         ];
-        $stats[] = ['Last 12 Month Hits', Db::fetchOne(
+        $stats[] = ['Last 12 Month Hits', (int) Db::fetchOne(
             "SELECT COUNT(*) FROM " . Common::prefixTable("log_link_visit_action") . " WHERE server_time >= ? AND server_time <= ?",
             [
                 Date::now()->setDay(1)->subMonth(12)->setTime('00:00:00')->toString('Y-m-d H:i:s'),
